PRS-313 Cleanup CPI subaccounts implementation
