+++ program-runtime/src/cpi.rs
@@ -18,7 +18,7 @@ use {
     solana_svm_measure::measure::Measure,
     solana_svm_timings::ExecuteTimings,
     solana_transaction_context::{
-        vm_slice::VmSlice, BorrowedInstructionAccount, IndexOfAccount, InstructionAccount,
+        vm_slice::VmSlice, BorrowedInstructionAccount, IndexOfAccount,
         MAX_ACCOUNTS_PER_INSTRUCTION, MAX_INSTRUCTION_DATA_LEN, SUBACCOUNT_MARKER,
     },
     std::mem,
@@ -518,19 +518,12 @@ pub trait SyscallInvokeSigned {
         check_aligned: bool,
     ) -> Result<Vec<TranslatedAccount<'a>>, Error>;
     /// F10: translate the caller-side subaccount AccountInfo array into
-    /// `TranslatedAccount` entries tagged with `SUBACCOUNT_MARKER`. The default
-    /// implementation returns an empty vector so existing Rust/C invoke
-    /// syscalls (which don't carry subaccounts) need no override. The
-    /// self-invoke syscalls override this to read `syscall_context.subaccounts_infos`.
+    /// `TranslatedAccount` entries tagged with `SUBACCOUNT_MARKER`.
     fn translate_subaccounts<'a>(
-        _subaccount_infos_addr: u64,
-        _subaccount_infos_len: u64,
-        _memory_mapping: &MemoryMapping<'_>,
-        _invoke_context: &mut InvokeContext,
-        _check_aligned: bool,
-    ) -> Result<Vec<TranslatedAccount<'a>>, Error> {
-        Ok(Vec::new())
-    }
+        memory_mapping: &MemoryMapping<'_>,
+        invoke_context: &mut InvokeContext,
+        check_aligned: bool,
+    ) -> Result<Vec<TranslatedAccount<'a>>, Error>;
     fn translate_signers(
         program_id: &Pubkey,
         signers_seeds_addr: u64,
@@ -821,7 +814,6 @@ pub fn cpi_common<S: SyscallInvokeSigned>(
     signers_seeds_addr: u64,
     signers_seeds_len: u64,
     memory_mapping: &mut MemoryMapping,
-    subaccounts: Vec<InstructionAccount>,
 ) -> Result<u64, Error> {
     // CPI entry.
     //
@@ -858,7 +850,7 @@ pub fn cpi_common<S: SyscallInvokeSigned>(
         check_aligned,
     )?;
     check_authorized_program(&instruction.program_id, &instruction.data, invoke_context)?;
-    invoke_context.prepare_next_instruction(instruction, &signers, subaccounts)?;
+    invoke_context.prepare_next_instruction(instruction, &signers)?;
 
     let mut accounts = S::translate_accounts(
         account_infos_addr,
@@ -868,20 +860,7 @@ pub fn cpi_common<S: SyscallInvokeSigned>(
         check_aligned,
     )?;
 
-    // F10: translate the caller-side subaccount AccountInfo slice, if any.
-    // `subaccounts_infos` is populated by `sol_set_subaccount_slice`; for
-    // regular invoke syscalls (non self-invoke) the default trait impl
-    // returns an empty vector.
-    let syscall_context = invoke_context.get_syscall_context()?;
-    let subaccounts_metadata_len = syscall_context.subaccounts_metadata.len() as u64;
-    let subaccount_infos_len = syscall_context.subaccounts_infos.vm_data_len;
-    let subaccount_infos_addr = syscall_context.subaccounts_infos.vm_data_addr;
-    if subaccounts_metadata_len != subaccount_infos_len {
-        return Err(Box::new(InstructionError::MissingAccount));
-    }
     let mut subaccounts = S::translate_subaccounts(
-        subaccount_infos_addr,
-        subaccount_infos_len,
         memory_mapping,
         invoke_context,
         check_aligned,
@@ -892,20 +871,15 @@ pub fn cpi_common<S: SyscallInvokeSigned>(
         // account (caller_account). We need to update the corresponding
         // BorrowedAccount (callee_account) so the callee can see the
         // changes. Subaccounts share the same flow, routed by the
-        // `SUBACCOUNT_MARKER` bit in `index_in_caller`. F10 slot
-        // subaccounts (loaded via `sol_load_subaccount`) carry an explicit
-        // `subaccount_slot = Some((tx_index, is_writable))` and use a
+        // `SUBACCOUNT_MARKER` bit in `index_in_caller` and use a
         // tx-level lookup since they are not in `instruction_subaccounts`.
         let transaction_context = &invoke_context.transaction_context;
         let instruction_context = transaction_context.get_current_instruction_context()?;
         for translated_account in accounts.iter_mut().chain(subaccounts.iter_mut()) {
-            let callee_account = if let Some((tx_idx, writable)) =
-                translated_account.subaccount_slot
-            {
-                instruction_context.try_borrow_subaccount_by_tx_index(tx_idx, writable)?
-            } else if translated_account.index_in_caller & SUBACCOUNT_MARKER != 0 {
-                let subaccount_index = translated_account.index_in_caller & !SUBACCOUNT_MARKER;
-                instruction_context.try_borrow_subaccount(subaccount_index)?
+            let callee_account = if translated_account.index_in_caller & SUBACCOUNT_MARKER != 0 {
+                instruction_context.try_borrow_subaccount_by_tx_index(
+                    translated_account.index_in_caller & !SUBACCOUNT_MARKER,
+                    translated_account.update_caller_account_info)?
             } else {
                 instruction_context
                     .try_borrow_instruction_account(translated_account.index_in_caller)?
@@ -935,17 +909,12 @@ pub fn cpi_common<S: SyscallInvokeSigned>(
     // CPI exit.
     //
     // Synchronize the callee's account changes so the caller can see them.
-    // Regular accounts and subaccounts share the same sync path; the
-    // `SUBACCOUNT_MARKER` high-bit in `index_in_caller` routes to the
-    // correct lane.
+    // Regular accounts and subaccounts share the same sync path.
     for translated_account in accounts.iter_mut().chain(subaccounts.iter_mut()) {
-        let mut callee_account = if let Some((tx_idx, writable)) =
-            translated_account.subaccount_slot
-        {
-            instruction_context.try_borrow_subaccount_by_tx_index(tx_idx, writable)?
-        } else if translated_account.index_in_caller & SUBACCOUNT_MARKER != 0 {
-            let subaccount_index = translated_account.index_in_caller & !SUBACCOUNT_MARKER;
-            instruction_context.try_borrow_subaccount(subaccount_index)?
+            let mut callee_account = if translated_account.index_in_caller & SUBACCOUNT_MARKER != 0 {
+                instruction_context.try_borrow_subaccount_by_tx_index(
+                    translated_account.index_in_caller & !SUBACCOUNT_MARKER,
+                    translated_account.update_caller_account_info)?
         } else {
             instruction_context
                 .try_borrow_instruction_account(translated_account.index_in_caller)?
@@ -965,13 +934,10 @@ pub fn cpi_common<S: SyscallInvokeSigned>(
 
     if stricter_abi_and_runtime_constraints {
         for translated_account in accounts.iter().chain(subaccounts.iter()) {
-            let mut callee_account = if let Some((tx_idx, writable)) =
-                translated_account.subaccount_slot
-            {
-                instruction_context.try_borrow_subaccount_by_tx_index(tx_idx, writable)?
-            } else if translated_account.index_in_caller & SUBACCOUNT_MARKER != 0 {
-                let subaccount_index = translated_account.index_in_caller & !SUBACCOUNT_MARKER;
-                instruction_context.try_borrow_subaccount(subaccount_index)?
+            let mut callee_account = if translated_account.index_in_caller & SUBACCOUNT_MARKER != 0 {
+                instruction_context.try_borrow_subaccount_by_tx_index(
+                    translated_account.index_in_caller & !SUBACCOUNT_MARKER,
+                    translated_account.update_caller_account_info)?
             } else {
                 instruction_context
                     .try_borrow_instruction_account(translated_account.index_in_caller)?
@@ -998,13 +964,6 @@ pub struct TranslatedAccount<'a> {
     pub caller_account: CallerAccount<'a>,
     pub update_caller_account_region: bool,
     pub update_caller_account_info: bool,
-    /// F10: when `Some((tx_index, is_writable))`, the entry refers to a
-    /// `sol_load_subaccount` slot rather than an instruction subaccount.
-    /// `index_in_caller` is unused for slot entries; CPI sync uses
-    /// `InstructionContext::try_borrow_subaccount_by_tx_index` against the
-    /// stored tx-level index + writability instead of the instruction-level
-    /// `try_borrow_subaccount` lookup.
-    pub subaccount_slot: Option<(IndexOfAccount, bool)>,
 }
 
 pub fn translate_account_infos<'a, T, F>(
@@ -1187,7 +1146,6 @@ where
                 caller_account,
                 update_caller_account_region: instruction_account.is_writable() || update_caller,
                 update_caller_account_info: instruction_account.is_writable(),
-                subaccount_slot: None,
             });
         } else {
             ic_msg!(
@@ -1200,95 +1158,6 @@ where
     }
 
     Ok(accounts)
-}
-
-/// F10: finish translating subaccounts from the caller's `AccountInfo`-slice
-/// view and build `TranslatedAccount` entries tagged with `SUBACCOUNT_MARKER`.
-///
-/// Parallel to [`translate_accounts_common`], but iterates the current
-/// instruction context's `instruction_subaccounts()` list and borrows each
-/// slot through `try_borrow_subaccount`. Sync with the callee is handled by
-/// the existing `cpi_common` pre/post-CPI loops via the `SUBACCOUNT_MARKER`
-/// dispatch — this helper does not call `update_callee_account` directly.
-pub fn translate_subaccounts_common<'a, T, F>(
-    subaccount_infos: &[T],
-    subaccount_infos_addr: u64,
-    invoke_context: &mut InvokeContext,
-    memory_mapping: &MemoryMapping<'_>,
-    check_aligned: bool,
-    do_translate: F,
-) -> Result<Vec<TranslatedAccount<'a>>, Error>
-where
-    F: Fn(
-        &InvokeContext,
-        &MemoryMapping<'_>,
-        bool,
-        u64,
-        &T,
-        &SerializedAccountMetadata,
-    ) -> Result<CallerAccount<'a>, Error>,
-{
-    let transaction_context = &invoke_context.transaction_context;
-    let instruction_context = transaction_context.get_current_instruction_context()?;
-
-    let subaccounts_metadata = &invoke_context
-        .get_syscall_context()
-        .unwrap()
-        .subaccounts_metadata;
-
-    let instruction_subaccounts = instruction_context.instruction_subaccounts();
-    let mut subaccounts = Vec::with_capacity(instruction_subaccounts.len());
-
-    for (index_in_caller, instruction_subaccount) in instruction_subaccounts.iter().enumerate() {
-        if instruction_context
-            .is_instruction_subaccount_duplicate(index_in_caller as IndexOfAccount)?
-            .is_some()
-        {
-            continue; // Skip duplicate subaccount
-        }
-
-        let serialized_metadata = subaccounts_metadata.get(index_in_caller).ok_or_else(|| {
-            ic_msg!(
-                invoke_context,
-                "Internal error: index mismatch for subaccount {}: out of bounds for \
-                 subaccounts_metadata with length {}",
-                index_in_caller,
-                subaccounts_metadata.len(),
-            );
-            Box::new(InstructionError::MissingAccount)
-        })?;
-        let subaccount_info = subaccount_infos.get(index_in_caller).ok_or_else(|| {
-            ic_msg!(
-                invoke_context,
-                "Internal error: invalid subaccount index {}: out of bounds for \
-                 subaccount_infos with length {}",
-                index_in_caller,
-                subaccount_infos.len(),
-            );
-            Box::new(InstructionError::MissingAccount)
-        })?;
-
-        let caller_account = do_translate(
-            invoke_context,
-            memory_mapping,
-            check_aligned,
-            subaccount_infos_addr.saturating_add(
-                (index_in_caller as u64).saturating_mul(mem::size_of::<T>() as u64),
-            ),
-            subaccount_info,
-            serialized_metadata,
-        )?;
-
-        subaccounts.push(TranslatedAccount {
-            index_in_caller: (index_in_caller as IndexOfAccount) | SUBACCOUNT_MARKER,
-            caller_account,
-            update_caller_account_region: instruction_subaccount.is_writable(),
-            update_caller_account_info: instruction_subaccount.is_writable(),
-            subaccount_slot: None,
-        });
-    }
-
-    Ok(subaccounts)
 }
 
 fn consume_compute_meter(invoke_context: &InvokeContext, amount: u64) -> Result<(), Error> {
@@ -1311,10 +1180,6 @@ fn consume_compute_meter(invoke_context: &InvokeContext, amount: u64) -> Result<
 /// `stricter_abi_and_runtime_constraints`) so the callee always sees the
 /// program's most recent header writes via the slot, mirroring what the
 /// standard CPI machinery does for declared subaccounts under stricter ABI.
-///
-/// Each returned entry is tagged with `subaccount_slot = Some((tx_idx,
-/// is_writable))`; `cpi_common`'s post-CPI sync uses that field to route
-/// state changes back through `try_borrow_subaccount_by_tx_index`.
 pub fn translate_subaccount_slots<'a, T, F>(
     invoke_context: &mut InvokeContext,
     memory_mapping: &MemoryMapping<'_>,
@@ -1399,7 +1264,6 @@ where
             caller_account,
             update_caller_account_region: is_writable,
             update_caller_account_info: is_writable,
-            subaccount_slot: Some((tx_idx, is_writable)),
         });
     }
 
@@ -1489,7 +1353,7 @@ fn update_callee_account(
     Ok(must_update_caller)
 }
 
-pub fn update_caller_account_region(
+fn update_caller_account_region(
     memory_mapping: &mut MemoryMapping,
     check_aligned: bool,
     caller_account: &CallerAccount,
@@ -1538,7 +1402,7 @@ pub fn update_caller_account_region(
 // Safety: Once `stricter_abi_and_runtime_constraints` is enabled all fields of [CallerAccount] used
 // in this function should never point inside the address space reserved for
 // accounts (regardless of the current size of an account).
-pub fn update_caller_account(
+fn update_caller_account(
     invoke_context: &InvokeContext,
     memory_mapping: &MemoryMapping<'_>,
     check_aligned: bool,
@@ -2199,11 +2063,8 @@ mod tests {
             .set_syscall_context(SyscallContext {
                 allocator: BpfAllocator::new(solana_program_entrypoint::HEAP_LENGTH as u64),
                 accounts_metadata: vec![account_metadata],
-                subaccounts_metadata: Vec::new(),
-                subaccounts_infos: crate::invoke_context::UntypedVmSlice::default(),
                 subaccount_slots: Vec::new(),
                 trace_log: Vec::new(),
-                dynamic_cpi_accounts: Vec::new(),
             })
             .unwrap();
