+++ programs/bpf_loader/src/lib.rs
@@ -24,7 +24,6 @@ use {
         execution_budget::MAX_INSTRUCTION_STACK_DEPTH,
         invoke_context::{
             BpfAllocator, InvokeContext, SerializedAccountMetadata, SubaccountSlot, SyscallContext,
-            UntypedVmSlice,
         },
         loaded_programs::{
             LoadProgramMetrics, ProgramCacheEntry, ProgramCacheEntryOwner, ProgramCacheEntryType,
@@ -264,7 +263,6 @@ fn create_vm<'a, 'b>(
     program: &'a Executable<InvokeContext<'b, 'b>>,
     regions: Vec<MemoryRegion>,
     accounts_metadata: Vec<SerializedAccountMetadata>,
-    subaccounts_metadata: Vec<SerializedAccountMetadata>,
     subaccount_slots: Vec<SubaccountSlot>,
     invoke_context: &'a mut InvokeContext<'b, 'b>,
     stack: &mut [u8],
@@ -286,14 +284,8 @@ fn create_vm<'a, 'b>(
     invoke_context.set_syscall_context(SyscallContext {
         allocator: BpfAllocator::new(heap_size as u64),
         accounts_metadata,
-        // F10: serialize_parameters populates subaccounts_metadata only when
-        // the instruction carries subaccounts (Wave 7+). For top-level invoke
-        // and regular CPIs the Vec is empty.
-        subaccounts_metadata,
-        subaccounts_infos: UntypedVmSlice::default(),
         subaccount_slots,
         trace_log: Vec::new(),
-        dynamic_cpi_accounts: Vec::new(),
     })?;
     Ok(EbpfVm::new(
         program.get_loader().clone(),
@@ -307,7 +299,7 @@ fn create_vm<'a, 'b>(
 /// Create the SBF virtual machine
 #[macro_export]
 macro_rules! create_vm {
-    ($vm:ident, $program:expr, $regions:expr, $accounts_metadata:expr, $subaccounts_metadata:expr, $subaccount_slots:expr, $invoke_context:expr $(,)?) => {
+    ($vm:ident, $program:expr, $regions:expr, $accounts_metadata:expr, $subaccount_slots:expr, $invoke_context:expr $(,)?) => {
         let invoke_context = &*$invoke_context;
         let stack_size = $program.get_config().stack_size();
         let heap_size = invoke_context.get_compute_budget().heap_size;
@@ -322,7 +314,6 @@ macro_rules! create_vm {
                 $program,
                 $regions,
                 $accounts_metadata,
-                $subaccounts_metadata,
                 $subaccount_slots,
                 $invoke_context,
                 stack
@@ -1499,7 +1490,6 @@ fn execute<'a, 'b: 'a>(
         parameter_bytes,
         regions,
         accounts_metadata,
-        subaccounts_metadata,
         subaccount_slots,
         instruction_data_offset,
     ) = serialization::serialize_parameters(
@@ -1511,12 +1501,9 @@ fn execute<'a, 'b: 'a>(
     serialize_time.stop();
 
     // save the account addresses so in case we hit an AccessViolation error we
-    // can map to a more specific error. F10: include the subaccount region so
-    // a violation inside a subaccount's VM range maps to the subaccount pubkey
-    // instead of falling through to a generic access-violation message.
+    // can map to a more specific error
     let account_region_addrs = accounts_metadata
         .iter()
-        .chain(subaccounts_metadata.iter())
         .map(|m| {
             let vm_end = m
                 .vm_data_addr
@@ -1530,11 +1517,6 @@ fn execute<'a, 'b: 'a>(
         })
         .collect::<Vec<_>>();
 
-    // F10: capture the main-account count before `accounts_metadata` is moved into
-    // `create_vm!` so the AccessViolation handler below can route subaccount-region
-    // indices (>= main count) through `try_borrow_subaccount` instead of
-    // unconditionally falling into the main lane (which produces `MissingAccount`).
-    let n_main_account_regions = accounts_metadata.len();
     let mut create_vm_time = Measure::start("create_vm");
     let execution_result = {
         let compute_meter_prev = invoke_context.get_remaining();
@@ -1543,7 +1525,6 @@ fn execute<'a, 'b: 'a>(
             executable,
             regions,
             accounts_metadata,
-            subaccounts_metadata,
             subaccount_slots,
             invoke_context
         );
@@ -1639,26 +1620,9 @@ fn execute<'a, 'b: 'a>(
                             let transaction_context = &invoke_context.transaction_context;
                             let instruction_context =
                                 transaction_context.get_current_instruction_context()?;
-                            // F10: `account_region_addrs` chains main accounts then
-                            // subaccounts. Indices `>= n_main_account_regions` belong
-                            // to the subaccount lane and must be borrowed through
-                            // `try_borrow_subaccount` — the original code unconditionally
-                            // routed to `try_borrow_instruction_account` which returns
-                            // `MissingAccount` for out-of-range indices, masking the real
-                            // access-violation classification (admin_update_oracle on
-                            // a previously-persisted subaccount surfaced as
-                            // `InstructionError::MissingAccount`).
-                            let account = if instruction_account_index < n_main_account_regions {
-                                instruction_context.try_borrow_instruction_account(
-                                    instruction_account_index as IndexOfAccount,
-                                )?
-                            } else {
-                                instruction_context.try_borrow_subaccount(
-                                    instruction_account_index
-                                        .saturating_sub(n_main_account_regions)
-                                        as IndexOfAccount,
-                                )?
-                            };
+                            let account = instruction_context.try_borrow_instruction_account(
+                                instruction_account_index as IndexOfAccount,
+                            )?;
                             if vm_addr.saturating_add(len) <= vm_addr_range.end {
                                 // The access was within the range of the accounts address space,
                                 // but it might not be within the range of the actual data.
@@ -1736,7 +1700,6 @@ fn execute<'a, 'b: 'a>(
             account_data_direct_mapping,
             parameter_bytes,
             &invoke_context.get_syscall_context()?.accounts_metadata,
-            &invoke_context.get_syscall_context()?.subaccounts_metadata,
         )
     }
