+++ syscalls/src/cpi.rs
@@ -3,9 +3,9 @@ use {
     solana_account_info::AccountInfo,
     solana_instruction::Instruction,
     solana_program_runtime::cpi::{
-        cpi_common, translate_account_infos, translate_accounts_c, translate_accounts_rust,
+        cpi_common, translate_accounts_c, translate_accounts_rust,
         translate_instruction_c, translate_instruction_rust, translate_signers_c,
-        translate_signers_rust, translate_subaccount_slots, translate_subaccounts_common,
+        translate_signers_rust, translate_subaccount_slots,
         CallerAccount, SolAccountInfo, SyscallInvokeSigned, TranslatedAccount,
     },
 };
@@ -30,7 +30,6 @@ declare_builtin_function!(
             signers_seeds_addr,
             signers_seeds_len,
             memory_mapping,
-            Vec::new(),
         )
     }
 );
@@ -62,40 +61,16 @@ impl SyscallInvokeSigned for SyscallInvokeSignedRust {
     }
 
     fn translate_subaccounts<'a>(
-        subaccount_infos_addr: u64,
-        subaccount_infos_len: u64,
         memory_mapping: &MemoryMapping<'_>,
         invoke_context: &mut InvokeContext,
         check_aligned: bool,
     ) -> Result<Vec<TranslatedAccount<'a>>, Error> {
-        let (subaccount_infos, _keys) = translate_account_infos(
-            subaccount_infos_addr,
-            subaccount_infos_len,
-            |subaccount_info: &AccountInfo| subaccount_info.key as *const _ as u64,
-            memory_mapping,
-            invoke_context,
-            check_aligned,
-        )?;
-        let mut subaccounts = translate_subaccounts_common(
-            subaccount_infos,
-            subaccount_infos_addr,
+        translate_subaccount_slots::<AccountInfo, _>(
             invoke_context,
             memory_mapping,
             check_aligned,
             CallerAccount::from_account_info,
-        )?;
-        // F10: append `sol_load_subaccount` slot entries so the CPI machinery
-        // syncs them across the call boundary. Pre-CPI sync (program view →
-        // AccountSharedData) runs inside `translate_subaccount_slots`; post-CPI
-        // sync runs through `cpi_common` via `subaccount_slot` dispatch.
-        let mut slot_entries = translate_subaccount_slots::<AccountInfo, _>(
-            invoke_context,
-            memory_mapping,
-            check_aligned,
-            CallerAccount::from_account_info,
-        )?;
-        subaccounts.append(&mut slot_entries);
-        Ok(subaccounts)
+        )
     }
 
     fn translate_signers(
@@ -135,7 +110,6 @@ declare_builtin_function!(
             signers_seeds_addr,
             signers_seeds_len,
             memory_mapping,
-            Vec::new(),
         )
     }
 );
@@ -167,38 +141,16 @@ impl SyscallInvokeSigned for SyscallInvokeSignedC {
     }
 
     fn translate_subaccounts<'a>(
-        subaccount_infos_addr: u64,
-        subaccount_infos_len: u64,
         memory_mapping: &MemoryMapping<'_>,
         invoke_context: &mut InvokeContext,
         check_aligned: bool,
     ) -> Result<Vec<TranslatedAccount<'a>>, Error> {
-        let (subaccount_infos, _keys) = translate_account_infos(
-            subaccount_infos_addr,
-            subaccount_infos_len,
-            |subaccount_info: &SolAccountInfo| subaccount_info.key_addr,
-            memory_mapping,
-            invoke_context,
-            check_aligned,
-        )?;
-        let mut subaccounts = translate_subaccounts_common(
-            subaccount_infos,
-            subaccount_infos_addr,
+        translate_subaccount_slots::<SolAccountInfo, _>(
             invoke_context,
             memory_mapping,
             check_aligned,
             CallerAccount::from_sol_account_info,
-        )?;
-        // F10: append `sol_load_subaccount` slot entries — see Rust impl
-        // above for the details.
-        let mut slot_entries = translate_subaccount_slots::<SolAccountInfo, _>(
-            invoke_context,
-            memory_mapping,
-            check_aligned,
-            CallerAccount::from_sol_account_info,
-        )?;
-        subaccounts.append(&mut slot_entries);
-        Ok(subaccounts)
+        )
     }
 
     fn translate_signers(
